Cleanup and factor out VoiceControlService
