+++ app/src/main/java/de/danoeh/antennapod/activity/MainActivity.java
@@ -9,8 +9,6 @@ import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.content.res.Configuration;
 import android.database.DataSetObserver;
-import android.media.session.PlaybackState;
-import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
@@ -33,16 +31,14 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ListView;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.bumptech.glide.Glide;
+import com.thumbjive.atennapod.VoiceControlService;
 
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.Validate;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.List;
 
 import de.danoeh.antennapod.R;
@@ -82,22 +78,12 @@ import rx.Subscription;
 import rx.android.schedulers.AndroidSchedulers;
 import rx.schedulers.Schedulers;
 
-import edu.cmu.pocketsphinx.Assets;
-import edu.cmu.pocketsphinx.Hypothesis;
-import edu.cmu.pocketsphinx.RecognitionListener;
-import edu.cmu.pocketsphinx.SpeechRecognizer;
-import edu.cmu.pocketsphinx.SpeechRecognizerSetup;
-
 import static android.widget.Toast.makeText;
-
-import de.danoeh.antennapod.core.service.playback.PlayerStatus;
-
 
 /**
  * The activity that is shown when the user launches the app.
  */
-public class MainActivity extends CastEnabledActivity implements NavDrawerActivity,
-        RecognitionListener {
+public class MainActivity extends CastEnabledActivity implements NavDrawerActivity {
 
     private static final String TAG = "MainActivity";
 
@@ -146,6 +132,8 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
     private ProgressDialog pd;
 
     private Subscription subscription;
+
+    private VoiceControlService voiceControlService;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -234,7 +222,8 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
 
         checkFirstLaunch();
 
-        runRecognizerSetup();
+        voiceControlService = new VoiceControlService(this);
+        voiceControlService.run();
     }
 
     private void saveLastNavFragment(String tag) {
@@ -823,226 +812,18 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
         setIntent(intent);
     }
 
-
     //
-    // Experimental voice integration
+    // VoiceControlService support
     //
 
-    private SpeechRecognizer recognizer;
-    private static final String KWS_SEARCH = "wakeup";
-//    private static final String KEYPHRASE = "hello voice";
-    private static final String KEYPHRASE = "wakeup";
-    private static final String COMMANDS_SEARCH = "commands";
-
-    private String currentSearch = KWS_SEARCH;
-
-    private void runRecognizerSetup() {
-        Log.i("TJ", "runRecognizerSetup");
-        // Recognizer initialization is a time-consuming and it involves IO,
-        // so we execute it in async task
-        new AsyncTask<Void, Void, Exception>() {
-            @Override
-            protected Exception doInBackground(Void... params) {
-                try {
-                    Log.i("TJ", "about to load assets");
-                    Assets assets = new Assets(MainActivity.this);
-                    File assetDir = assets.syncAssets();
-                    Log.i("TJ", "after syncAssets");
-                    setupRecognizer(assetDir);
-                } catch (IOException e) {
-                    Log.i("TJ", "asset load error: " + e);
-                    return e;
-                }
-                return null;
-            }
-
-            @Override
-            protected void onPostExecute(Exception result) {
-//                showToast("Speech recognizer initialized.\nSay \"" + KEYPHRASE + "\" to wake up");
-                showToast("Speech recognizer initialized");
-                if (result != null) {
-//                    ((TextView) findViewById(R.id.caption_text))
-//                            .setText("Failed to init recognizer " + result);
-                } else {
-                    switchSearch(KWS_SEARCH);
-//                    switchSearch(COMMANDS_SEARCH);
-                 }
-            }
-        }.execute();
-    }
-
-    private void switchSearch(String searchName) {
-        Log.i("TJ", "switchSearch: " + searchName);
-        currentSearch = searchName;
-
-        if (currentSearch.startsWith(KWS_SEARCH)) {
-            showToast("Say \"" + KEYPHRASE + "\" to wake up");
-        } else if (currentSearch.startsWith(COMMANDS_SEARCH)) {
-            showToast("Say \"play\", \"pause\", or \"sleep\"");
-        }
-
-
-//        String caption = getResources().getString(captions.get(currentSearch));
-//        ((TextView) findViewById(R.id.caption_text)).setText(caption);
-//
-//        if (currentSearch.equals(COMMANDS_SEARCH)) {
-//            resetState();
-//        }
-
-        restartRecognizer();
-    }
-
-    private void restartRecognizer() {
-        Log.i("TJ", "restartRecognizer, currentSearch: " + currentSearch);
-        recognizer.stop();
-
-        // If we are not spotting, start listening with timeout (10000 ms or 10 seconds).
-        if (currentSearch.equals(KWS_SEARCH))
-            recognizer.startListening(currentSearch);
-        else
-            recognizer.startListening(currentSearch, 30000);
-    }
-
-    private void setupRecognizer(File assetsDir) throws IOException {
-        // The recognizer can be configured to perform multiple searches
-        // of different kind and switch between them
-
-        recognizer = SpeechRecognizerSetup.defaultSetup()
-                .setAcousticModel(new File(assetsDir, "en-us-ptm"))
-                .setDictionary(new File(assetsDir, "cmudict-en-us.dict"))
-
-                .setRawLogDir(assetsDir) // To disable logging of raw audio comment out this call (takes a lot of space on the device)
-
-                .getRecognizer();
-        Log.i("TJ", "recognizer instantiated");
-        recognizer.addListener(this);
-
-        // Create keyword-activation search.
-        recognizer.addKeyphraseSearch(KWS_SEARCH, KEYPHRASE);
-
-        // Hello Voice POC commands
-        File commandsGrammar = new File(assetsDir, "commands.gram");
-        recognizer.addGrammarSearch(COMMANDS_SEARCH, commandsGrammar);
-        Log.i("TJ", "grammer added");
-    }
-
-
-    /**
-     * In partial result we get quick updates about current hypothesis. In
-     * keyword spotting mode we can react here, in other modes we need to wait
-     * for final result in onResult.
-     */
-    @Override
-    public void onPartialResult(Hypothesis hypothesis) {
-        if (hypothesis == null)
-            return;
-
-        String text = hypothesis.getHypstr();
-        Log.i("TJ", "hypothesis: " + text);
-//        showToast("onPartial: " + text);
-//        ((TextView) findViewById(R.id.result_text)).setText(text);
-    }
-
-    /**
-     * This callback is called when we stop the recognizer.
-     */
-    @Override
-    public void onResult(Hypothesis hypothesis) {
-//        ((TextView) findViewById(R.id.result_text)).setText("");
-        if (hypothesis != null) {
-            String text = hypothesis.getHypstr();
-            Log.i("TJ", "onResult: " + text);
-//            showToast("onResult: " + text);
-            if (text.equals(KEYPHRASE)) {
-                switchSearch(COMMANDS_SEARCH);
-//                resetState();
-            }
-//
-//            if ("exit".equals(text) || "quit".equals(text)) {
-//                showToast("exiting");
-//                switchSearch(KWS_SEARCH);
-//            } if ("cancel".equals(text)) {
-//                showToast("resetting state");
-//                resetState();
-            if ("play".equals(text)) {
-                handlePlayCommand();
-            } else if ("pause".equals(text)) {
-                handlePauseCommand();
-            } else if ("sleep".equals(text)) {
-                switchSearch(KWS_SEARCH);
-            }
-//            } else if (text.startsWith("go ")) {
-//                handleGoCommand(text);
-//            } else if (text.equals("faster") || text.equals("slower")) {
-//                handleSpeedCommand(text);
-//            } else if (text.contains("session")) {
-//                handleSessionCommand(text);
-//            } else if (text.contains("feedback")) {
-//                handleFeedbackCommand(text);
-//            } else  if (text.equals("help")) {
-//                showToast("don't forget your towel");
-//            } else {
-//                showToast(text);
-//            }
-        }
-//        displayState();
-    }
-
-    private void handlePlayCommand() {
-        PlayerStatus status = getPlaybackController().getStatus();
-        Log.i("TJ", "handlePlay - status: " + status);
-        if (status != PlayerStatus.PLAYING) {
-            Log.i("TJ", "playing...");
-//            showToast("playing");
-            getPlaybackController().playPause();
-        }
-    }
-
-    private void handlePauseCommand() {
-        PlayerStatus status = getPlaybackController().getStatus();
-        Log.i("TJ", "handlePlay - status: " + status);
-        if (status == PlayerStatus.PLAYING) {
-            Log.i("TJ", "pausing...");
-//            showToast("pausing");
-            getPlaybackController().playPause();
-        }
-    }
-
-    @Override
-    public void onBeginningOfSpeech() {
-    }
-
-    /**
-     * We stop recognizer here to get a final result
-     */
-    @Override
-    public void onEndOfSpeech() {
-        Log.i("TJ", "onEndOfSpeech");
-        restartRecognizer();
-    }
-
-    private void showToast(String text) {
+    public void showToast(String text) {
         if (text != null && text.length() > 0) {
             makeText(getApplicationContext(), text, Toast.LENGTH_SHORT).show();
         }
     }
 
-    @Override
-    public void onError(Exception error) {
-        Log.e("TJ", "onError: " + error.getMessage());
-//        ((TextView) findViewById(R.id.caption_text)).setText(error.getMessage());
-        showToast(error.getMessage());
-    }
-
-    @Override
-    public void onTimeout() {
-        switchSearch(KWS_SEARCH);
-    }
-
-
     public PlaybackController getPlaybackController() {
         ExternalPlayerFragment fragment = (ExternalPlayerFragment) getSupportFragmentManager().findFragmentByTag(ExternalPlayerFragment.TAG);
         return fragment.getPlaybackControllerTestingOnly();
     }
-
 }
